requirements.txt
Flask==2.0.1
PyPDF2==2.1.0
python-docx==0.8.11